Move Convert<CrewPosition> to script/crewPosition.h
Move template from script/enum.h to script/crewPosition.h and
remove the related include. This fixes a strict compilation issue
on MSVC and consolidates the two related templates to one place.
