+++ src/script/crewPosition.h
@@ -1,11 +1,25 @@
 #pragma once
 
 #include "../crewPosition.h"
-#include "script/enum.h"
 #include "script/environment.h"
 
-
 namespace sp::script {
+template<> struct Convert<CrewPosition> {
+    static int toLua(lua_State* L, CrewPosition value) {
+        lua_pushstring(L, crewPositionToString(value).c_str());
+        return 1;
+    }
+    static CrewPosition fromLua(lua_State* L, int idx) {
+        string str = string(luaL_checkstring(L, idx)).lower();
+
+        auto result = tryParseCrewPosition(str);
+        if (!result.has_value()) {
+            luaL_error(L, "Unknown CrewPosition: %s", str.c_str());
+        }
+        return result.value_or(CrewPosition::helmsOfficer);
+    }
+};
+
 template<> struct Convert<CrewPositions> {
     static int toLua(lua_State* L, CrewPositions value) {
         lua_newtable(L);
